ref(eslint): Convert all packages to use central eslint config (#4111)
Convert nextjs, react, node, and serverless to using the central eslint config. This should make sure all packages use a consistent eslint config. In our central eslint config, I've disabled our no-async-await rule in tests and made jsdoc required for only exported methods/classes. 

For browser we've kept ignorePatterns: ['test/integration/**', 'src/loader.js'],. I think it's fine to not lint the loader, and the integration tests are going to be refactored with the release stability work, so we can keep this for now. 

Note: Ember is not converted as it's special cased (part of the build) - so I kept it separate. We can have a discussion about what we wanna do with ember, but my opinion is that it's fine to keep it a special case.