fix(nextjs): Show full stacktraces in integration tests (#4125)
By default, Node only displays the top 10 frames in an error's stacktrace. Especially with async code (which, when down-compiled, introduces lots of frames just with its async-i-ness), this means you can be missing a lot of frames. This removes that cap for our integration tests, so we can see the full stacktrace when debugging.