feat(awslambda) : Capture errors individually on sqs partial batch failure (#4130)
A lambda can receive up to 1000 messages. (if we use a batchWindow), but it can fail because of 10/1000 messages failed. Tools such as [Middy](https://middy.js.org/) (and the `sqs-partial-batch-failure` plugin) allow managing this case by deleting records from the queue individually, so that we don't have to retry a record that has already worked.

This PR suggests adding an option to capture all errors coming from a `Promise.allSettled`. The `wrapHandler` will not fail the lambda even if there are errors, because I assume that it will be the middy’s job.

Co-authored-by: J4YM <jeremy@modjo.ai>