ci: Run browser integration tests also on Firefox and Webkit (#4232)
We configure a job matrix to run browser tests in parallel, even though
Karma could connect to all browsers at once. The reason for this is
that, if run time is not a concern (less than 2min at the moment),
running each browser as a separate job seemed more robust and easier to
spot browser-specific problems when a test fails.

We get all browser binaries from Playwright, so we don't need to rely on
versions installed on the host system. (And this is also the easiest way
to get a headless Webkit browser on Ubuntu)

We include the Playwright browsers in cache to avoid having to download
them again over and over.

We install Webkit system dependencies on every run of the Webkit tests
because it is inconvenient to cache and restore system dependencies.
Installing then is safer, and perhaps even faster, than dealing with the
cache.