ref(utils): Change addInstrumentationHandler to take reg args
Passing in an object is not necessary as there is always going to be two
arguments required. We can just convert this to explicitly passing the
options, it saves on bundle size.

In addition, we can remove the guards around `type` and `callback`
because addInstrumentationHandler is an internally used function, so
Typescript and our test suite should handle the function getting called
with arguments of the correct type.