feat(core): Add processing metadata to scope and event (#4252)
There are various situations where it's helpful to be able to provide data to the event processing pipeline, without that data necessarily being sent to Sentry. 

Rather than piggyback that data in `debug_meta`, and then have to do a "is there still anything in it, or should I delete it" dance after that data is pulled back out, it would simpler if we just had a spot where we could chuck anything we wanted with the knowledge that all of the data there will get filtered out of the final event.

This introduces such a spot, called `sdkProcessingMetadata`, both on the scope (so that the data can get set at any point prior to the event getting captured) and in the event. No use is (yet) made of this - that will come in future PRs.

Note: While there is a public setter for this data on the scope, I'm thinking of this as an internal API, and therefore am not planning to document it for users.