ref(tracing): Remove legacy method of getting transaction from scope (#4394)
In https://github.com/getsentry/sentry-javascript/pull/2952, a new way of pulling the currently-active transaction off of the scope was introduced. That new way is no longer so new, and the legacy way was left in place at that time only out of an (over-)abundance of caution. It's time we removed it, both for ease of reading and for the few bytes we'll gain in bundle size reduction.