ref(core): Use `sdkProcessingMetadata` field for transaction sampling data (#4388)
In https://github.com/getsentry/sentry-javascript/pull/4252, a new `sdkProcessingMetadata` field was added to the `Scope` and `Event` interfaces, to provide a place to put data which is needed for in-SDK event processing but shouldn't be sent to Sentry. This refactors one step in that processing... process... to use the new field, eliminating the need to pull the data out of `debug_meta` before the event is sent.