feat: Undeprecate Severity Enum (#4412)
In https://github.com/getsentry/sentry-javascript/pull/4280,
specifically in commit dd3aa70e5edb8efc317cb1ced7cf87e614e3957e, we
deprecated the `Severity` enum in favour of using a string union type,
`SeverityLevel`. It's important to note that this change affected the
type signature of one of our public API methods, `captureMessage`.

The change to deprecate the `Severity` enum was done for bundle size
reasons.

After releasing the beta with these changes, it was found that
deprecating the `Severity` enum and replacing it with `SeverityLevel`
was quite the disruptive change, which would make upgrading to the minor
version a hassle for users. As a result, this patch undeprecates the
`Severity` enum. The `Severity` enum will be removed in the upcoming
major release instead, as a breaking change.