fix(nextjs): Inject `sentry.x.config.js` into `pages/_error` (#4397)
There are certain situations where `_error` is loaded before `_app` has had a chance to run. In those cases, we miss errors caught in `_error`, because the SDK has not yet initialized (which currently happens the first time `_app` runs). This injects the initialization code into `_error` as well, to avoid such problems.

(In order to not add extra code to the browser bundle, and because this has only shown up as a problem on the server side, for the moment this only does the injection when building the server bundle.)

Fixes https://github.com/getsentry/sentry-javascript/issues/4168.