ref(serverless): Remove `rethrowAfterCapture` use in AWS lambda wrapper (#4448)
This PR removes the use of the `rethrowAfterCapture` option from the AWS handler-wrapping function `@sentry/serverless` provides. Without its use, the code follows the same path as it has been when the option has been set to its default value, which is `true`. Though this is _technically_ a breaking change for anyone who currently has it set to `false`, 

a) the option has never been documented,
b) it never should have existed in the first place (we do error _monitoring_, not error _handling_, meaning there should be no question as to whether we rethrow the error after recording it),
c) of all of the public projects on GH, only three are using it set explicitly to `true`, and zero are using it set to `false`.

For the three projects using it set to `true`, there is no behavior change. Further, since all three are written in TS, the option hasn't been removed from the `WrapperOptions` type, so that their continuing to pass it doesn’t cause type errors. It has been made optional, though, so that our non-use of it also doesn't cause type errors. There's also a TODO so that we remember to remove it in v7.

Fixes https://github.com/getsentry/sentry-javascript/issues/3364.