fix(nextjs): Revert #4139 - remove manipulation of `res.finished` value (#4516)
In https://github.com/getsentry/sentry-javascript/pull/4139, a change was introduced in order to suppress a false positive warning thrown by nextjs, by setting `res.finished` to `true` just long enough for nextjs to check it and decide no warning was needed. In simple cases, it worked just fine, but in some set-ups the "just long enough for nextjs to check it and calm down" turned out to also be "just long enough for other things to check it and get mad."

This backs out that change, as it seems it's doing more harm than good. In order to address the original problem (documented in [1] and [2]), a new warning is added, explaining that the false positive is just that. Not as elegant as suppressing the message altogether, but it should tide us over until such time as we're able to try again with a different approach.

Fixes https://github.com/getsentry/sentry-javascript/issues/4151.

[1] https://github.com/getsentry/sentry-javascript/issues/3852
[2] https://github.com/getsentry/sentry-javascript/issues/4007