fix(utils): Make new non-enumerable properties mutable (#4528)
There are a few places in the SDK where we add properties to an object in order to store data for internal use. In cases where there's a chance that object might be serialized, we need to make sure the new properties are non-enumerable, which we do with the (appropriately-named) function `addNonEnumerableProperty`.

Currently, the non-enumerable properties we add are also immutable, even though nothing requires them to be, simply because that's the default when using `Object.defineProperty()`. This blocks use-cases in which a user might choose to modify an internal property's value in order to change how the SDK behaves. (See the attached issue for an example of such a use-case.)

This PR explicitly makes these properties mutable. It also changes the docstring to reflect the fact that the function can be used for more than just setting properties on function objects.

Fixes https://github.com/getsentry/sentry-javascript/issues/4525.