fix(tracing): Export `BrowserTracing` directly in CDN bundle (#4570)
When we made the changes in https://github.com/getsentry/sentry-javascript/pull/4204 (to export `BrowserTracing` directly rather than as part of `Sentry.Integrations`), we didn't make the change in the tracing CDN bundle. While It's true that there's no bundle size issue there (mongo and friends never were getting included in the CDN bundle, so there was nothing to fix), it did mean that when we changed the docs to match the above PR's changes, we ended up breaking CDN folks' usage of `BrowserTracing`.

This fixes that by making the same change in the tracing CDN bundle as was made for the npm package.

Fixes https://github.com/getsentry/sentry-docs/issues/4722.