chore(build): Update and hoist `rollup` (#4578)
As prep for the new bundling process, this updates `rollup` to latest and hoists it to the main `package.json`.

Other than single-character changes to generated variable names, the only effect this has on the generated bundles is to add `Object.defineProperty(exports, '__esModule', { value: true });` (or its minified counterpart) at the end of each bundle. According to the Rollup docs[1]:

> This property signifies that the exported value is the namespace of an ES module and that the default export of this module corresponds to the .default property of the exported object.

Since we don't have default exports from our SDKs or integrations, this isn't necessary and just adds dead weight to the bundles. Setting `output.esModule` to `false` in the Rollup config prevents this addition.

[1] https://rollupjs.org/guide/en/#outputesmodule