chore(build): Remove unused instances of `@rollup/plugin-commonjs` (#4590)
In four of the five packages in which we're using `@rollup/plugin-commonjs`, it has no effect (removing it doesn't change the bundles at all). This makes sense, because nowhere in those four packages are we importing any CJS modules. This removes those extraneous plugin instances.

(The exception is `@sentry/integrations`, because the `Offline` plugin imports `localforage`, which is CJS. As a result, its rollup config hasn't been changed.)