chore(build): Use standard regex for wasm terser `mangle.properties` option (#4592)
Currently, `@sentry/wasm`'s minification settings (the `mangle.properties` option in rollup's terser plugin) prevent the mangling of any property names. (Note: This is different from variable names, which do get mangled.)

In other packages, internal property names (those starting with a single `_`) _are_ mangled. This ports that setting to the wasm package, which turns out not to meaningfully change the bundle, because nowhere in the package do such properties appear. As a result, the only effects of this change are different single-letter variables being used in the final minified bundle. (For a screenshot of the changes, see the corresponding PR.)