ref(build): Centralize identical CDN rollup config (#4637)
There is a great deal of overlap between the rollup config files we use to generate our CDN bundles. As part of moving to the new bundling process, this pulls the most obvious pieces of shared config into a central file. (In this case, “most obvious” means that they are strictly identical across files, with no tweaks needed. A follow-up PR will handle the places where there are slight differences between one rollup config and another.)

As one would hope when moving things without changing them, there is no effect on the content of the bundles produced.